src/requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.6
cryptography==35.0.0
defusedxml==0.7.1
Django==3.2.9
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==3.0.1
django-cleanup==5.2.0
django-taggit==1.5.1
django-tinymce4-lite==1.8.0
easy-thumbnails==2.7.1
idna==3.2
jsmin==3.0.0
oauthlib==3.1.1
Pillow==8.3.2
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==2.2.0
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
soupsieve==2.2.1
sqlparse==0.4.2
Unidecode==1.3.2
urllib3==1.26.7
